# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GenAff API Gateway
  version: 1.0.0
  description: |
    **GenAff** is a multi-provider AI API gateway with per-request billing in INR paisa,
    wallet top-ups via Razorpay, API-key management, and rate limiting.

    ## Authentication
    Two security schemes are in use:

    | Scheme | Header | Used by |
    |--------|--------|---------|
    | `BearerAuth` | `Authorization: Bearer <JWT>` | Dashboard, admin, key management |
    | `ApiKeyAuth` | `Authorization: Bearer <USER_API_KEY>` | Proxy endpoints (`/v1/*`) |

    ## Billing Overview
    1. New users receive **free units** (requests or tokens, configurable).
    2. Once free units are exhausted, each request deducts cost from the wallet in **INR paisa** (1 paisa = ₹0.01).
    3. Cost is estimated **before** the provider call; over/under-charges are reconciled after.
    4. If wallet balance is insufficient, a **402 Payment Required** is returned.

    ## Rate Limits
    Default limits per API key:
    - **20 requests / minute**
    - **500 requests / day**

    Rate-limit headers are returned on every proxy response:
    ```
    X-RateLimit-Minute-Remaining: 19
    X-RateLimit-Day-Remaining: 499
    ```

  contact:
    name: GenAff Support
    email: shouriyatayal1234@gmail.com
  license:
    name: MIT
servers:
  - url: https://genaff-api.shauryacodes.xyz
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtained from `POST /auth/login`.
        Payload: `{ sub: userId, email, role, iat, exp }`.
        Pass as: `Authorization: Bearer <token>`
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        User-generated API key (prefixed `sk_`).
        Obtained from `POST /keys` — shown **once** at creation.
        Pass as: `Authorization: Bearer sk_<key>`

  # ─── Reusable Schemas ──────────────────────────────────────────────────────
  schemas:
...
    Error500:
      type: object
      properties:
        error: { type: string, example: "Internal server error" }
...
paths:
...
  /webhooks/payments:
    post:
      tags: [Webhooks]
      summary: Razorpay payment webhook
      description: |
        Receives `payment.captured` and `order.paid` events from Razorpay and credits
        the matching `TopUpTransaction` to the user's wallet.

        ### Signature Verification
        Every webhook request from Razorpay includes an HMAC-SHA256 signature:
        ```
        x-razorpay-signature: <hex_digest>
        ```
        The digest is computed over the **raw request body** using `RAZORPAY_WEBHOOK_SECRET`.

        #### Verification (Node.js)
        ```js
        import crypto from 'crypto';

        function verify(rawBody, signature, secret) {
          const expected = crypto
            .createHmac('sha256', secret)
            .update(rawBody)
            .digest('hex');
          return crypto.timingSafeEqual(
            Buffer.from(expected, 'hex'),
            Buffer.from(signature, 'hex')
          );
        }
        ```

        ### Event: `payment.captured`
        ```json
        {
          "event": "payment.captured",
          "payload": {
            "payment": {
              "entity": {
                "id": "pay_AbcXyz123",
                "order_id": "order_AbcXyz123",
                "receipt": "<TopUpTransaction.id>",
                "amount": 10000,
                "currency": "INR",
                "status": "captured"
              }
            }
          }
        }
        ```
        The `receipt` field maps to the `TopUpTransaction.id` stored in the database.

        ### Event: `order.paid`
        Similar structure with `payload.order.entity.receipt`.

        ### Idempotency
        If the transaction is already `completed`, the webhook is acknowledged (`200 OK`) without re-crediting.

        Unhandled events are always acknowledged with `{ "received": true, "event": "<name>" }`.
      parameters:
        - name: x-razorpay-signature
          in: header
          required: true
          schema: { type: string }
          description: HMAC-SHA256(rawBody, RAZORPAY_WEBHOOK_SECRET) in hex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string, example: "payment.captured" }
                payload: { type: object }
            example:
              event: "payment.captured"
              payload:
                payment:
                  entity:
                    id: "pay_AbcXyz123"
                    receipt: "tx-uuid-1234"
                    amount: 10000
                    currency: "INR"
                    status: "captured"
      responses:
        '200':
          description: Webhook acknowledged (always returned on valid events)
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean, example: true }
                  event: { type: string, example: "payment.captured" }
        '400':
          description: Invalid signature or malformed body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Invalid signature" }
        '500':
          description: Webhook processing error (Razor retries on 5xx)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error500' }